+++ src/ts/modules/Settings.ts
@@ -4,6 +4,7 @@ import { DOM } from './DOM'
 import { centerPanel } from './CenterPanel'
 import { InitSettings } from './InitSettings'
 import { SettingsForm } from './Settings.d'
+import { BeautifyForm } from './BeautifyForm'
 
 // 设置表单
 class Settings {
@@ -13,6 +14,8 @@ class Settings {
     this.bindEvents()
 
     new InitSettings(this.form)
+
+    new BeautifyForm(this.form)
   }
 
   public form: SettingsForm
@@ -38,29 +41,28 @@ class Settings {
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_下载作品类型的提示Center'
   )}">${lang.transl('_下载作品类型')}<span class="gray1"> ? </span></span>
-  <input type="checkbox" name="downType0" id="setWorkType0" checked><label for="setWorkType0"> ${lang.transl(
-    '_插画'
-  )}&nbsp;</label>
-  <input type="checkbox" name="downType1" id="setWorkType1" checked><label for="setWorkType1"> ${lang.transl(
-    '_漫画'
-  )}&nbsp;</label>
-  <input type="checkbox" name="downType2" id="setWorkType2" checked><label for="setWorkType2"> ${lang.transl(
-    '_动图'
-  )}&nbsp;</label>
+  <input type="checkbox" name="downType0" id="setWorkType0" class="need_beautify checkbox_common" checked>
+  <span class="beautify_checkbox"></span>
+  <label for="setWorkType0"> ${lang.transl('_插画')}&nbsp;</label>
+  <input type="checkbox" name="downType1" id="setWorkType1" class="need_beautify checkbox_common" checked>
+  <span class="beautify_checkbox"></span>
+  <label for="setWorkType1"> ${lang.transl('_漫画')}&nbsp;</label>
+  <input type="checkbox" name="downType2" id="setWorkType2" class="need_beautify checkbox_common" checked>
+  <span class="beautify_checkbox"></span>
+  <label for="setWorkType2"> ${lang.transl('_动图')}&nbsp;</label>
   </p>
   <p class="${this.optionClass}" data-no="3">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_怎样下载多图作品'
   )}">${lang.transl('_多图作品设置')}<span class="gray1"> ? </span></span>
-  <input type="radio" name="multipleImageWorks" id="multipleImageWorks1" value="0">
-  <label for="multipleImageWorks1"> ${lang.transl(
-    '_全部下载'
-  )}&nbsp; </label>
-  <input type="radio" name="multipleImageWorks" id="multipleImageWorks2" value="-1">
-  <label for="multipleImageWorks2"> ${lang.transl(
-    '_不下载'
-  )}&nbsp; </label>
-  <input type="radio" name="multipleImageWorks" id="multipleImageWorks3" value="1">
+  <input type="radio" name="multipleImageWorks" id="multipleImageWorks1" class="need_beautify radio" value="0">
+  <span class="beautify_radio"></span>
+  <label for="multipleImageWorks1"> ${lang.transl('_全部下载')}&nbsp; </label>
+  <input type="radio" name="multipleImageWorks" id="multipleImageWorks2" class="need_beautify radio" value="-1">
+  <span class="beautify_radio"></span>
+  <label for="multipleImageWorks2"> ${lang.transl('_不下载')}&nbsp; </label>
+  <input type="radio" name="multipleImageWorks" id="multipleImageWorks3" class="need_beautify radio" value="1">
+  <span class="beautify_radio"></span>
   <label for="multipleImageWorks3"> ${lang.transl(
     '_下载前几张图片'
   )}&nbsp; </label>
@@ -70,40 +72,34 @@ class Settings {
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_动图保存格式title'
   )}">${lang.transl('_动图保存格式')}<span class="gray1"> ? </span></span>
-  <input type="radio" name="ugoiraSaveAs" id="ugoiraSaveAs1" value="webm" checked>
-  <label for="ugoiraSaveAs1"> ${lang.transl(
-    '_webmVideo'
-  )} &nbsp;</label>
-  <input type="radio" name="ugoiraSaveAs" id="ugoiraSaveAs3" value="gif"> 
-  <label for="ugoiraSaveAs3">${lang.transl(
-    '_gif'
-  )} &nbsp;</label>
-  <input type="radio" name="ugoiraSaveAs" id="ugoiraSaveAs2" value="zip"> 
-  <label for="ugoiraSaveAs2">${lang.transl(
-    '_zipFile'
-  )} &nbsp;</label>
+  <input type="radio" name="ugoiraSaveAs" id="ugoiraSaveAs1" class="need_beautify radio" value="webm" checked>
+  <span class="beautify_radio"></span>
+  <label for="ugoiraSaveAs1"> ${lang.transl('_webmVideo')} &nbsp;</label>
+  <input type="radio" name="ugoiraSaveAs" id="ugoiraSaveAs3" class="need_beautify radio" value="gif"> 
+  <span class="beautify_radio"></span>
+  <label for="ugoiraSaveAs3">${lang.transl('_gif')} &nbsp;</label>
+  <input type="radio" name="ugoiraSaveAs" id="ugoiraSaveAs2" class="need_beautify radio" value="zip"> 
+  <span class="beautify_radio"></span>
+  <label for="ugoiraSaveAs2">${lang.transl('_zipFile')} &nbsp;</label>
   </p>
   <p class="${this.optionClass}" data-no="5">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_筛选收藏数的提示Center'
   )}">${lang.transl('_筛选收藏数Center')}<span class="gray1"> ? </span></span>
-  <input type="radio" name="checkFavNum" id="checkFavNum0" value="0" checked>
-  <label for="checkFavNum0">  ${lang.transl(
-    '_不限制'
-  )}&nbsp; </label>
-  <input type="radio" name="checkFavNum" id="checkFavNum1" value="1">
-  <label for="checkFavNum1">  ${lang.transl(
-    '_大于'
-  )}&nbsp; </label>
+  <input type="radio" name="checkFavNum" id="checkFavNum0" class="need_beautify radio" value="0" checked>
+  <span class="beautify_radio"></span>
+  <label for="checkFavNum0">  ${lang.transl('_不限制')}&nbsp; </label>
+  <input type="radio" name="checkFavNum" id="checkFavNum1" class="need_beautify radio" value="1">
+  <span class="beautify_radio"></span>
+  <label for="checkFavNum1">  ${lang.transl('_大于')}&nbsp; </label>
   <input type="text" name="setFavNum" class="setinput_style1 blue" value="0">&nbsp;&nbsp;&nbsp;&nbsp;
   </p>
   <p class="${this.optionClass}" data-no="6">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_只下载已收藏的提示'
   )}">${lang.transl('_只下载已收藏')}<span class="gray1"> ? </span></span>
-  <input type="checkbox" name="setOnlyBmk" id="setOnlyBmk"> <label for="setOnlyBmk">${lang.transl(
-    '_启用'
-  )}</label>
+  <input type="checkbox" name="setOnlyBmk" id="setOnlyBmk" class="need_beautify checkbox_switch"> 
+  <span class="beautify_switch"></span>
   </p>
   <p class="${this.optionClass}" data-no="7">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
@@ -112,58 +108,55 @@ class Settings {
     '_筛选宽高的按钮文字'
   )}<span class="gray1"> ? </span></span>
   <input type="text" name="setWidth" class="setinput_style1 blue" value="0">
-  <input type="radio" name="setWidthAndOr" id="setWidth_AndOr1" value="&" checked> <label for="setWidth_AndOr1">and&nbsp;</label>
-  <input type="radio" name="setWidthAndOr" id="setWidth_AndOr2" value="|"> <label for="setWidth_AndOr2">or&nbsp;</label>
+  <input type="radio" name="setWidthAndOr" id="setWidth_AndOr1" class="need_beautify radio" value="&" checked>
+  <span class="beautify_radio"></span>
+  <label for="setWidth_AndOr1">and&nbsp;</label>
+  <input type="radio" name="setWidthAndOr" id="setWidth_AndOr2" class="need_beautify radio" value="|">
+  <span class="beautify_radio"></span>
+  <label for="setWidth_AndOr2">or&nbsp;</label>
   <input type="text" name="setHeight" class="setinput_style1 blue" value="0">
   </p>
   <p class="${this.optionClass}" data-no="8">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_设置宽高比例Title'
   )}">${lang.transl('_设置宽高比例')}<span class="gray1"> ? </span></span>
-  <input type="radio" name="ratio" id="ratio0" value="0" checked> 
-  <label for="ratio0"> ${lang.transl(
-    '_不限制'
-  )}&nbsp; </label>
-  <input type="radio" name="ratio" id="ratio1" value="1">
-  <label for="ratio1"> ${lang.transl(
-    '_横图'
-  )}&nbsp; </label>
-  <input type="radio" name="ratio" id="ratio2" value="2">
-  <label for="ratio2"> ${lang.transl(
-    '_竖图'
-  )}&nbsp; </label>
-  <input type="radio" name="ratio" id="ratio3" value="3">
-  <label for="ratio3"> ${lang.transl(
-    '_输入宽高比'
-  )}</label>
+  <input type="radio" name="ratio" id="ratio0" class="need_beautify radio" value="0" checked> 
+  <span class="beautify_radio"></span>
+  <label for="ratio0"> ${lang.transl('_不限制')}&nbsp; </label>
+  <input type="radio" name="ratio" id="ratio1" class="need_beautify radio" value="1">
+  <span class="beautify_radio"></span>
+  <label for="ratio1"> ${lang.transl('_横图')}&nbsp; </label>
+  <input type="radio" name="ratio" id="ratio2" class="need_beautify radio" value="2">
+  <span class="beautify_radio"></span>
+  <label for="ratio2"> ${lang.transl('_竖图')}&nbsp; </label>
+  <input type="radio" name="ratio" id="ratio3" class="need_beautify radio" value="3">
+  <span class="beautify_radio"></span>
+  <label for="ratio3"> ${lang.transl('_输入宽高比')}</label>
   <input type="text" name="userRatio" class="setinput_style1 blue" value="1.4">
   </p>
   <p class="${this.optionClass}" data-no="9">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_设置id范围提示'
-  )}">${lang.transl('_设置id范围')}&nbsp;&nbsp; <span class="gray1"> ? </span></span>
-  <input type="radio" name="idRange" id="idRange0" value="0" checked>
-  <label for="idRange0">  ${lang.transl(
-    '_不限制'
-  )}&nbsp; </label>
-  <input type="radio" name="idRange" id="idRange1" value="1">
-  <label for="idRange1">  ${lang.transl(
-    '_大于'
-  )}&nbsp; </label>
-  <input type="radio" name="idRange" id="idRange2" value="2">
-  <label for="idRange2">  ${lang.transl(
-    '_小于'
-  )}&nbsp; </label>
+  )}">${lang.transl(
+    '_设置id范围'
+  )}&nbsp;&nbsp; <span class="gray1"> ? </span></span>
+  <input type="radio" name="idRange" id="idRange0" class="need_beautify radio" value="0" checked>
+  <span class="beautify_radio"></span>
+  <label for="idRange0">  ${lang.transl('_不限制')}&nbsp; </label>
+  <input type="radio" name="idRange" id="idRange1" class="need_beautify radio" value="1">
+  <span class="beautify_radio"></span>
+  <label for="idRange1">  ${lang.transl('_大于')}&nbsp; </label>
+  <input type="radio" name="idRange" id="idRange2" class="need_beautify radio" value="2">
+  <span class="beautify_radio"></span>
+  <label for="idRange2">  ${lang.transl('_小于')}&nbsp; </label>
   <input type="text" name="idRangeInput" class="setinput_style1 w100 blue" value="">
   </p>
   <p class="${this.optionClass}" data-no="10">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_设置投稿时间提示'
   )}">${lang.transl('_设置投稿时间')} <span class="gray1"> ? </span></span>
-  <input type="checkbox" name="postDate" id="setPostDate">
-  <label for="setPostDate"> ${lang.transl(
-    '_启用'
-  )}</label>
+  <input type="checkbox" name="postDate" id="setPostDate" class="need_beautify checkbox_switch">
+  <span class="beautify_switch"></span>
   <input type="datetime-local" name="postDateStart" placeholder="yyyy-MM-dd HH:mm" class="setinput_style1 postDate blue" value="">
   &nbsp;-&nbsp;
   <input type="datetime-local" name="postDateEnd" placeholder="yyyy-MM-dd HH:mm" class="setinput_style1 postDate blue" value="">
@@ -268,10 +261,8 @@ class Settings {
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_添加字段名称提示'
   )}">${lang.transl('_添加字段名称')}<span class="gray1"> ? </span></span>
-  <input type="checkbox" name="tagNameToFileName" id="setTagNameToFileName" checked>
-  <label for="setTagNameToFileName"> ${lang.transl(
-    '_启用'
-  )}</label>
+  <input type="checkbox" name="tagNameToFileName" id="setTagNameToFileName" class="need_beautify checkbox_switch" checked>
+  <span class="beautify_switch"></span>
   &nbsp;&nbsp;&nbsp;
   <span class="showFileNameResult"> ${lang.transl('_预览文件名')}</span>
   </p>
@@ -279,10 +270,8 @@ class Settings {
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_快速下载建立文件夹提示'
   )}">${lang.transl('_快速下载建立文件夹')}<span class="gray1"> ? </span></span>
-  <input type="checkbox" name="alwaysFolder" id="setAlwaysFolder" >
-  <label for="setAlwaysFolder"> ${lang.transl(
-    '_启用'
-  )}</label>
+  <input type="checkbox" name="alwaysFolder" id="setAlwaysFolder" class="need_beautify checkbox_switch" >
+  <span class="beautify_switch"></span>
   </p>
   <p class="${this.optionClass}" data-no="16">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
@@ -294,19 +283,15 @@ class Settings {
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_快速下载的提示'
   )}">${lang.transl('_是否自动下载')}<span class="gray1"> ? </span></span>
-  <input type="checkbox" name="quietDownload" id="setQuietDownload" checked>
-  <label for="setQuietDownload"> ${lang.transl(
-    '_启用'
-  )}</label>
+  <input type="checkbox" name="quietDownload" id="setQuietDownload" class="need_beautify checkbox_switch" checked>
+  <span class="beautify_switch"></span>
   </p>
   <p class="${this.optionClass}" data-no="18">
   <span class="has_tip settingNameStyle1" data-tip="${lang.transl(
     '_预览搜索结果说明'
   )}">${lang.transl('_预览搜索结果')}<span class="gray1"> ? </span></span>
-  <input type="checkbox" name="previewResult" id="setPreviewResult" checked>
-  <label for="setPreviewResult"> ${lang.transl(
-    '_启用'
-  )}</label>
+  <input type="checkbox" name="previewResult" id="setPreviewResult" class="need_beautify checkbox_switch" checked>
+  <span class="beautify_switch"></span>
   </p>
   <input type="hidden" name="debut" value="0">
   </form>`